Making whole project works

+++ Java_StartupHorror_Simulator/src/com/company/project/Project.java
@@ -19,8 +19,8 @@ public class Project {
     public int price;
     public int penalty;
     public Map.Entry<ProjectComplexity, Integer> complexity;
-    public int chancesToBugs = 0;
-    public int chancesToDelay = 0;
+    public double chancesToBugs = 0;
+    public double chancesToDelay = 0;
 
 
     public Project(Calendar currentDate) {
